+++ NEWS.md
@@ -166,3 +166,9 @@
 * Fixed a bug that sometimes caused an error in the estimation algorithm because the random degrees of freedom parameters were too close to their strict lower bound 2. This change may have an effect on the estimation results of StMVAR and G-StMVAR models with specific seed.
 * Adjusted the tolerance when the estimation functions warns about near-non-stationary estimates or near-singular error term covariance matrices.
 * Updated the vignette.
+
+# gmvarkit 2.0.3
+
+* Fixed a bug that caused an error when trying to estimate model a with restrictions on the mean parameters or a G-StMVAR model by setting up initial population in the genetic algorithm.
+* Added new functionality to the function gsmvar_to_sgsmvar: now it is possible to employ Cholesky identification if M=1.
+* Added new function: estimate_sgsmvar which can be used to conveniently estimate overidentified structural models.